🔒 Nevermind Laravel Gates

+++ app/Http/Controllers/GameController.php
@@ -126,6 +126,11 @@ class GameController extends Controller
             return response()->json(['message' => 'Game not found'], 404);
         }
 
+        if ($request->user()->id !== $game->manager && $request->user()->role !== 'admin') {
+            return response()->json(['message' => 'Unauthorized'], 403);
+        }
+        
+
         $request->validate([
             'name' => 'string|max:255',
             'description' => 'string',
@@ -163,9 +168,16 @@ class GameController extends Controller
     {
         $game = Game::find($id);
 
-        $game->delete();
-        
-        return response()->json(['message' => 'Game deleted successfully'], 200);
+        if (!$game) {
+            return response()->json(['message' => 'Game not found'], 404);
+        }
+
+        if ($game->manager === auth()->user()->id || auth()->user()->role === 'admin') {
+            $game->delete();
+            return response()->json(['message' => 'Game deleted successfully'], 200);
+        }
+
+        return response()->json(['message' => 'Unauthorized'], 403);
     }
 
     public function storeGenre(Request $request)
